requirements.txt
discord
gpt4all
python-dotenv